DHoughFind.cc DTrackCandidate_factory_CDC.cc DTrackCandidate_factory_FDC.cc   - Fixed a bunch of warnings due to not explicitly casting doubles when assigning them to integer types
DTrackCandidate_factory_FDCCathodes.cc
  - Removed references to fields no longer in DTrackCandidate class. These members were redendant with some inherited from DKinematicData

+++ libraries/TRACKING/DHoughFind.cc
@@ -238,8 +238,8 @@ void DHoughFind::ClearPoints(void)
 //---------------------------------
 void DHoughFind::FindIndexes(const DVector2 &pos, int &ix, int &iy)
 {
-	ix = floor((pos.X()-xmin)/bin_widthx);
-	iy = floor((pos.Y()-ymin)/bin_widthy);
+	ix = (int)floor((pos.X()-xmin)/bin_widthx);
+	iy = (int)floor((pos.Y()-ymin)/bin_widthy);
 }
 
 //---------------------------------
